Dev: split countryUrl tests into node and jsdom
